+++ src/components/ArticleForm/ArticleSetupModal/CategorySetup/CategoryList/index.tsx
@@ -6,7 +6,7 @@ import { loadCategories } from './_shared'
 import { SelectedCategoryType } from '../../_shared'
 
 import classNames from 'classnames/bind'
-import styles from './CategoryList.module.css'
+import styles from './index.module.css'
 
 const cx = classNames.bind(styles)
